+++ Server/Server/Server/translate.h
@@ -3,8 +3,8 @@ class CTranslate
 {
 public:
 	CTranslate( );
-	void Translate( long Address, const char* Text, ... );
+	void Translate( int Address, const char* Text, ... );
 	void SetTexts( );
 private:
-	static long m_Address_Text, m_Spacements;
+	static int m_Address_Text, m_Spacements;
 };
